==> Rmd files/1-tidy_data.Rmd <==
---
title: "Tidying Data"
author: "Nicole Bednarik & Katie Thomas"
date: 8 August 2022
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---


```{r setup, include = FALSE}

# Load package libraries
library(plyr)
library(kableExtra)
library(gridExtra)
library(knitr)
library(tidyverse)

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)
```

# Data

## Morphological measurements

We recorded 6 morphological measurements from each salamander specimen: snout-vent length, wet mass, corneal diameters (left and right), and transverse external eye diameters (left and right). Symmetrical measures (e.g. eye diameters) were averaged for each individual prior to further analysis. 

```{r load-data}

# Load raw data for measured frog specimens
salamanders_morph <- data.frame(read.csv("../Data/Raw/salamanders_analysis.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))

# Tidy dataset for analysis
salamanders.subset <- salamanders_morph %>%
  mutate(rootmass = Mass_g^(1/3)) %>% #adds cube root of mass
  mutate(eyemean = rowMeans(salamanders_morph[c('ED_right_mm', 'ED_left_mm')], na.rm=TRUE)) %>% #adds mean of L/R eyes
  mutate(cormean = rowMeans(salamanders_morph[c('CD_right_mm', 'CD_left_mm')], na.rm=TRUE)) %>%  #adds mean of L/R corneas
  mutate(eyemean = na_if(eyemean, "NaN")) %>% #remove missing values
  mutate(cormean = na_if(cormean, "NaN")) %>% #remove missing values
  select(Order, Suborder, Family, Subfamily, Genus, Species, Genus_Species, SVL_mm, Mass_g, rootmass, ED_right_mm, ED_left_mm, eyemean, CD_right_mm, CD_left_mm, cormean, Gill_Presence) #keeps only columns of interest for analyses
```

## Ecological traits

We coded 6 ecological traits for each species from published literature into binary categories, and then merged this data with our specimen measurements by genus and species. 

Note that for one species (_Pleurodeles waltl_), both a gilled and an ungilled specimen were measured. The ungilled specimen is labelled as Pleurodeles_waltl, and the gilled as Pleurodeles_waltl_gilled in the Genus_Species column. 

```{r trait-data, results = "hide"}

#Import ecological trait data with blank cells reading as NA
traits <- data.frame(read.csv("../Data/Raw/salamanders_traits.csv", na.strings = c("", " "))) 

#keep only relevent columns for merging
traits.subset <- traits %>%
  select(-one_of(c("X","Order","Suborder","Family","Subfamily","Genus","Species")))

```


## Merged data

Here, we merge the morphological data with the trait data. We then export the merged, tidied dataset to use with our analyses. Because the two morphs of P. waltl are coded in the species column, we can merge by species.

```{r}
# Merge morphology dataset with ecological trait dataset
salamanders <- merge(salamanders.subset, traits.subset, by="Genus_Species", 
                   all.x = TRUE, all.y = FALSE)

# Check structure of final dataframe for analyses (numbers and factors in proper columns)
str(salamanders)

# Export tidied dataset for analyses
write.csv(salamanders, file = "../Data/Tidy/salamanders_tidy.csv")
```

## Sampling

We sampled all 9 extant caudata families, and morphological sampling for interspecific analysis ranged from n = 1 to n = 6 individuals for each species (based on availability in the collections).

```{r sampling, results = "hide"}

#Number of species and individuals sampled 
counts <-ddply(salamanders, .(salamanders$Family, salamanders$Subfamily, salamanders$Genus, salamanders$Species), nrow)
names(counts) <- c("Family","Subfamily", "Genus","Species","Sampled")

#create scrolling RMarkdown table of sampling
kable(counts[ , c("Family","Genus","Species","Sampled")], caption = "Species and sampling effort for morphological data from museum specimens.") %>%
  kable_styling(full_width = F) %>%
  collapse_rows(columns = 1, valign = "top") %>%
  scroll_box(height = "500px")

#export sampling data for supplement
write.csv(counts, file = "../Data/Tidy/sampling.csv")

```
