==> Rmd files/5-boxplots.new.rmd <==
---
title: "Salamander eye scaling and investment - boxplots"
author: "Nicole Bednarik & Katie Thomas"
date: 9 September 2022
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---

```{r setup, include = FALSE}

# Load package libraries
library(plyr)
library(geiger)
library(caper)
library(plotly)
library(kableExtra)
library(gridExtra)
library(ggimage)
library(grid)
library(gtable)
library(cowplot)
library(knitr)
library(tidyverse)

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)

```

# Data

First we import our tidied salamander morphology & trait binary dataset with the residuals of a cornea ~ cube root of mass PGLS as a column.

```{r load-data}

# Import tidied salamander eye size and trait data
sal.res <- data.frame(read.csv("../Data/salamanders_residuals_binary.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))
sal.res.2 <- data.frame(read.csv("../Data/salamanders_residuals_updated.csv", header=TRUE, na.strings=c("", "NA", " ", "<1")))

# Quick look at data structure
str(sal.res)
str(sal.res.2)

sal.res <- sal.res[-c(35:460)]

```
#Plots

Next, we make boxplots of how eye size and relative eye size (represented by the PGLS residuals)compare across salamander ecological traits.

##Aquatic

```{r}
#boxplot of cornea diameter vs aquatic ecology
aquatic.abs <- ggplot(data = filter(sal.res, !is.na(Aquatic)), aes(x = Aquatic, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Aquatic ecology") +
  theme(legend.position = "none")

#print plot
aquatic.abs

```
```{r}
#boxplot of pgls residuals vs. aquatic ecology (relative eye size)
aquatic.rel <- ggplot(data = filter(sal.res, !is.na(Aquatic)), aes(x = Aquatic, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Aquatic ecology") +
  theme(legend.position = "none")

#print plot
aquatic.rel

```

##Semiaquatic
```{r}
#boxplot of cornea diameter vs semiaquatic ecology
semiaquatic.abs <- ggplot(data = filter(sal.res, !is.na(Semiaquatic)), aes(x = Semiaquatic, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Semiaquatic ecology") +
  theme(legend.position = "none")

#print plot
semiaquatic.abs

```
```{r}
#boxplot of pgls residuals vs. semiaquatic ecology (relative eye size)
semiaquatic.rel <- ggplot(data = filter(sal.res, !is.na(Semiaquatic)), aes(x = Semiaquatic, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Semiaquatic ecology") +
  theme(legend.position = "none")

#print plot
semiaquatic.rel

```

##Terrestrial
```{r}
#boxplot of cornea diameter vs terrestrial ecology
terrestrial.abs <- ggplot(data = filter(sal.res, !is.na(Terrestrial)), aes(x = Terrestrial, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Terrestrial ecology") +
  theme(legend.position = "none")

#print plot
terrestrial.abs

```

```{r}
#boxplot of pgls residuals vs. terrestrial ecology (relative eye size)
terrestrial.rel <- ggplot(data = filter(sal.res, !is.na(Terrestrial)), aes(x = Terrestrial, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Terrestrial ecology") +
  theme(legend.position = "none")

#print plot
terrestrial.rel

```

##Scansorial
```{r}
#boxplot of cornea diameter vs scansorial ecology
scansorial.abs <- ggplot(data = filter(sal.res, !is.na(Scansorial)), aes(x = Scansorial, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Scansorial ecology") +
  theme(legend.position = "none")

#print plot
scansorial.abs

```

```{r}
#boxplot of pgls residuals vs. scansorial ecology (relative eye size)
scansorial.rel <- ggplot(data = filter(sal.res, !is.na(Scansorial)), aes(x = Scansorial, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Scansorial ecology") +
  theme(legend.position = "none")

#print plot
scansorial.rel

```


##Subfossorial
```{r}
#boxplot of cornea diameter vs subfossorial ecology
subfossorial.abs <- ggplot(data = filter(sal.res, !is.na(Subfossorial)), aes(x = Subfossorial, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Subfossorial ecology") +
  theme(legend.position = "none")

#print plot
subfossorial.abs

```

```{r}
#boxplot of pgls residuals vs. subfossorial ecology (relative eye size)
subfossorial.rel <- ggplot(data = filter(sal.res, !is.na(Subfossorial)), aes(x = Subfossorial, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Subfossorial ecology") +
  theme(legend.position = "none")

#print plot
subfossorial.rel

```

##Cave_dwelling
```{r}
#boxplot of cornea diameter vs cave dwelling ecology
cave_dwelling.abs <- ggplot(data = filter(sal.res, !is.na(Cave_dwelling)), aes(x = Cave_dwelling, y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Cave dwelling ecology") +
  theme(legend.position = "none")

#print plot
cave_dwelling.abs

```

```{r}
#boxplot of pgls residuals vs. cave dwelling ecology (relative eye size)
cave_dwelling.rel <- ggplot(data = filter(sal.res, !is.na(Cave_dwelling)), aes(x = Cave_dwelling, y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Cave dwelling ecology") +
  theme(legend.position = "none")

#print plot
cave_dwelling.rel
```

##Eye size and investment across adult habitats
```{r}
#boxplot of cornea diameter vs adult habitat
adult_habitat.abs <- ggplot(data = filter(sal.res.2, !is.na(Adult.Habitat)), aes(x = Adult.Habitat, y = cor_av, col=Adult.Habitat)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Adult Habitat") +
  theme(legend.position = "none")

#print plot
adult_habitat.abs

```
```{r}
#boxplot of pgls residuals vs. adult habitat (relative eye size)
adult_habitat.rel <- ggplot(data = filter(sal.res.2, !is.na(Adult.Habitat)), aes(x = Adult.Habitat, y = pglsres.cdmass, col=Adult.Habitat)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Adult Habitat") +
  theme(legend.position = "none")

#print plot
adult_habitat.rel
```

##Eye size and investment relationship with development type

```{r}
#boxplot of cornea diameter vs development
development.abs <- ggplot(data = filter(sal.res.2, !is.na(Development)), aes(x = Development, y = cor_av, col=Development)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Mean cornea diameter (mm)") +
  xlab("Development") +
  theme(legend.position = "none")

#print plot
development.abs

```

```{r}
#boxplot of pgls residuals vs. development (relative eye size)
development.rel <- ggplot(data = filter(sal.res.2, !is.na(Development)), aes(x = Development, y = pglsres.cdmass, col=Development)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA) +
  stat_summary(fun.y = mean, colour="black", geom="point", 
               shape=18, size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species), 
              shape = 19, size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea diameter (PGLS residuals)") +
  xlab("Development") +
  theme(legend.position = "none")

#print plot
development.rel
```

# Ecology statistics

The habitat category looks the most interesting to me, so here we run a PGLS for eye-body allometry with habitat as a covariate to see if habitat has an effect on relative eye size in salamanders. 

```{r}

# Prep data and tree for analysis------

#import phylogeny
caudatatree <- read.nexus("../Data/caudata-tree")

#Make row names of dataset the species names (so it will match phylogeny tips)
rownames(sal.res.2) <- sal.res.2$Genus_Species

#check that names match in dataframe and tree
name.check(phy = caudatatree, data = sal.res.2)
drops <- name.check(phy = caudatatree, data = sal.res.2)

#drop the blind salamander from the phylogeny for plots
tree.hab <- drop.tip(caudatatree, drops$tree_not_data)

#use caper function to combine phylogeny and data into one object (this function also matches species names in tree and dataset)
sal.comp <- comparative.data(phy = tree.hab, data = sal.res.2, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)

#check for dropped tips or dropped species
sal.comp$dropped$tips #phylogeny
sal.comp$dropped$unmatched.rows #dataset

# Run habitat pgls--------

#PGLS cornea diameter vs. cube root of mass + habitat
pgls_hab <- pgls(log10(cor_av) ~ log10(rootmass_av) + Adult.Habitat,
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_hab)
par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.hab <- pgls.profile(pgls_hab, "lambda")
plot(lambda.hab)

#main effects
anova(pgls_hab)

#print model output 
summary(pgls_hab)
```

# Color  figures with pallettes in the frog paper

Here I will just re-run the code to make the figures you want for your poster, with a custom color pallette so that the habitat categories will be the same for all of them. 

First we define what we want that color pallette to be:

```{r}
#see what the levels of adult habitat are named
levels(as.factor(sal.res.2$Adult.Habitat))

#define a colorblind-friendly vector of colors for adult habitat
col_hab <- c("aquatic" = "#0072B2",
             "scansorial" = "#009E73",
             "semiaquatic" = "#56B4E9",
             "subfossorial" = "#CC79A7",
             "terrestrial" = "#E69F00")

#now you can see that each habitat is assigned a hex color when you look at the vector
col_hab

#make a vector of point shapes for adult habitat
shape_hab <- c("aquatic" = 25,
             "scansorial" = 24,
             "semiaquatic" = 22,
             "subfossorial" = 21,
             "terrestrial" = 23)

#see that each state has a shape now
shape_hab
```

## Boxplot adult habitat figure

You've already made a nice boxplot of adult habitats and relative eye size, so now we can just modify that plot to make it nicer, ordered sensibly, and colored how we want. 

```{r}

#boxplot of pgls residuals vs. adult habitat (relative eye size)
adult_habitat.rel <- ggplot(data = filter(sal.res.2, !is.na(Adult.Habitat)), aes(x = reorder(Adult.Habitat, pglsres.cdmass, fun = mean), y = pglsres.cdmass, fill=Adult.Habitat, shape = Adult.Habitat)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha = 0.5) +
  stat_summary(fun.y = mean, colour="black", geom="point", size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, shape = Adult.Habitat),  size = 1.3, alpha = 0.3, col = "black", fill = "black", position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea size") +
  xlab("Adult habitat") +
  theme(legend.position = "none") +
# use our custom colors for habitat
  scale_fill_manual(values = col_hab, name = "Adult habitat") +
# use our custom shapes for points
  scale_shape_manual(values = shape_hab, name = "Adult habitat")

#print plot
adult_habitat.rel

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(adult_habitat.rel)

#export plot to pdf - you can change the export dimensions by changing the height and width numbers to make it fit well on your poster!
pdf(file = "boxplots.pdf", height=4, width=6)
adult_habitat.rel
dev.off()

```

## Eye body scaling figure

Here we re-make the allometry plot for cornea vs. body mass and color the points by adult habitat. 

```{r}

#PGLS cornea diameter vs. cube root of mass
pgls_cdmass <- pgls(log10(cor_av) ~ log10(rootmass_av),
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#print model output 
summary(pgls_cdmass)

#plot
cdmass.plot <- ggplot(data = sal.res.2, aes(x = rootmass_av, y = cor_av, col = Adult.Habitat, fill = Adult.Habitat, shape = Adult.Habitat, text = Genus_Species)) +
 geom_point(alpha = 0.7) +
  #make x scale log10
  scale_x_log10(name = "Cube root of mass (g)") +
  #make y scale log10
  scale_y_log10(name = "Cornea diameter (mm)") +
  # use our custom colors for habitat
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  scale_fill_manual(values = col_hab, name = "Adult habitat") +
  # use our custom shapes for habitat
  scale_shape_manual(values = shape_hab, name = "Adult habitat") +
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  geom_abline(slope = coef(pgls_cdmass)[[2]], intercept = coef(pgls_cdmass)[[1]]) 

#print plot
cdmass.plot

#interactive plot
ggplotly(cdmass.plot)

#export plot to pdf - you can change the export dimensions by changing the height and width numbers to make it fit well on your poster!
pdf(file = "allometry.pdf", height=4, width=6)
cdmass.plot
dev.off()
```

## Phylogeny figures

Here we re-make the phylogeny figure, but color the bars for absolute and relative eye size by adult habitat to mirror the other figures for the poster

```{r}
#make dataframe for labeling phylogeny figure
sal.phy <- sal.res.2 %>%
  #make column labeling tip labels that match phylogeny
  mutate(tip = Genus_Species) %>%
  #put genus and species in separate columns
  separate(Genus_Species, c("genus", "species"), sep = "_", extra = "drop") %>%
  #add tip labels with regular text
  mutate(labels = as.factor(paste(genus, species, sep = " "))) %>%
  #make hab column with adult habitat
  mutate(hab = Adult.Habitat) %>%
  #keep only columns we will use for plotting
  select(tip, labels, cor_av, pglsres.cdmass, hab)

# set row names in dataset to match the tip labels in the tree
row.names(sal.phy) <- sal.phy$tip

#check that phylogeny and data match exactly
name.check(tree.hab, sal.phy)

#resort  dataset to the order of tree tip labels
sal.phy <- sal.phy[tree.hab$tip.label, ] 

#labels
labs <- sal.phy %>% select(tip, labels)
```

Next, we use ggtree to plot the phylogeny alongside bar plots for absolute cornea diameter and relative conea size. 

```{r, fig.height=18, fig.width=6}

library(ggtree)

# Make the phylogeny plot
p <- ggtree(tree.hab) %<+% labs + 
     geom_tiplab(size = 3, aes(label = labels), fontface = 3) + 
   xlim_tree(375) + 
   coord_cartesian(clip = 'off') 

# Make a second plot next to the phylogeny (bars for cornea diameter)
p2 <- facet_plot(p, panel="Cornea diameter (mm)", data=sal.phy, geom=geom_segment, aes(x=0, xend=cor_av, y=y, yend=y, color = hab), size = 3, alpha = 0.7) +
  scale_color_manual(values = col_hab,
                     name = "Adult habitat")

# Make a third plot next to the first two (bars for relative cornea size)
p3 <- facet_plot(p2, panel='Relative cornea size', data=sal.phy, geom=geom_segment, aes(x=0, xend=pglsres.cdmass, y=y, yend=y, color = hab), size=3, alpha = 0.7) +
  theme_tree2(legend.position=c(.07, .93)) +# add scale bars and move legends
  scale_color_manual(values = col_hab,
                     name = "Adult habitat")

#control widths of panels
 gt = ggplot_gtable(ggplot_build(p3))
# gtable_show_layout(gt) # will show you the layout - very handy function
# gt # see plot layout in table format
# gt$layout$l[grep('tree', gt$layout$name)] # you want to find the column specific to panel-2
 gt$widths[7] = 0.5*gt$widths[5] # reduce column 7
 gt$widths[9] = 0.5*gt$widths[5]

#print figure
plot(gt)

#export figure
pdf("tree-figure.pdf", width = 8, height = 22)
grid::grid.draw(gt) # plot with grid draw
dev.off()
```
